fix(ground_truth): deduplicate shared edges in mesh_geodesic_eikonal
Interior edges were being added twice (once per adjacent triangle),
causing csr_matrix to sum duplicate weights and double edge costs.
Also adds wall_nodes(), make_phi_true_wall(), ECG lead debug tools,
and non-identifiability mock (reflection symmetry, y=0.5 electrodes).

+++ fisiocomPinn/Grid.py
@@ -203,9 +203,9 @@ class Grid:
 
         if isinstance(q, torch.Tensor):
             dev  = q.device
-            f0t  = torch.tensor(f[:, 0], device=dev)
-            f1t  = torch.tensor(f[:, 1], device=dev)
-            f2t  = torch.tensor(f[:, 2], device=dev)
+            f0t  = torch.tensor(f[:, 0], dtype=torch.long, device=dev)
+            f1t  = torch.tensor(f[:, 1], dtype=torch.long, device=dev)
+            f2t  = torch.tensor(f[:, 2], dtype=torch.long, device=dev)
             At   = torch.tensor(self._areas, dtype=q.dtype, device=dev)
             Gt   = torch.tensor(self._grads, dtype=q.dtype, device=dev)
